added sync async dry-run

+++ api-tool/src/main/java/jp/chang/myclinic/apitool/Main.java
@@ -3,11 +3,16 @@ package jp.chang.myclinic.apitool;
 import com.github.javaparser.StaticJavaParser;
 import com.github.javaparser.ast.CompilationUnit;
 import com.github.javaparser.ast.Modifier.Keyword;
+import com.github.javaparser.ast.NodeList;
 import com.github.javaparser.ast.body.CallableDeclaration.Signature;
 import com.github.javaparser.ast.body.ClassOrInterfaceDeclaration;
 import com.github.javaparser.ast.body.MethodDeclaration;
 import com.github.javaparser.ast.body.Parameter;
+import com.github.javaparser.ast.expr.SimpleName;
 import com.github.javaparser.ast.stmt.BlockStmt;
+import com.github.javaparser.ast.type.ClassOrInterfaceType;
+import com.github.javaparser.ast.type.PrimitiveType;
+import com.github.javaparser.ast.type.Type;
 
 import java.io.File;
 import java.io.IOException;
@@ -17,7 +22,7 @@ import java.nio.file.Paths;
 import java.util.*;
 import java.util.stream.Collectors;
 
-import static java.util.stream.Collectors.toSet;
+import static java.util.stream.Collectors.*;
 
 public class Main {
 
@@ -25,7 +30,10 @@ public class Main {
     private Path backendPersistDir = backendDir.resolve("persistence");
     private Path backendMockDir = Paths.get("backend-mock/src/main/java/jp/chang/myclinic/backendmock");
     private Path backendPersistMockDir = backendMockDir.resolve("persistence");
+    private Path asyncDir = Paths.get("backend-async/src/main/java/jp/chang/myclinic/backendasync");
     private Set<String> persists;
+    private List<MethodDeclaration> backendMethods;
+    private boolean dryRun = false;
 
     public static void main(String[] args) throws Exception {
         new Main().run(args);
@@ -33,7 +41,58 @@ public class Main {
 
     private void run(String[] args) throws Exception {
         this.persists = listPersists(backendPersistDir);
-        syncMock();
+        this.dryRun = true;
+        CompilationUnit backendUnit = StaticJavaParser.parse(backendDir.resolve("Backend.java"));
+        ClassOrInterfaceDeclaration backendClass = backendUnit.getClassByName("Backend").orElseThrow(
+                () -> new RuntimeException("Cannot find Backend class")
+        );
+        backendMethods = backendClass.getMethods().stream()
+                .filter(m ->
+                    m.isPublic() && !m.isAnnotationPresent("BackendPrivate")
+                )
+                .collect(toList());
+        //syncMock();
+        syncAsync();
+    }
+
+    private void syncAsync() {
+        Map<Signature, MethodDeclaration> backendSigs = methodsToSigMap(backendMethods);
+        CompilationUnit asyncUnit = parseSource(asyncDir.resolve("BackendAsync.java"));
+        ClassOrInterfaceDeclaration asyncInterface = getInterface(asyncUnit, "BackendAsync");
+        Map<Signature, MethodDeclaration> asyncSigs = methodsToSigMap(asyncInterface.getMethods());
+        Set<Signature> missing = findMissingSigs(asyncSigs.keySet(), backendSigs.keySet());
+        for(Signature sig: missing){
+            MethodDeclaration backendMethod = backendSigs.get(sig);
+            MethodDeclaration asyncMethod = asyncInterface.addMethod(backendMethod.getNameAsString());
+            asyncMethod.setType(makeAsyncReturnType(backendMethod.getType()));
+            asyncMethod.removeBody();
+            for(Parameter param: backendMethod.getParameters()){
+                asyncMethod.addParameter(param);
+            }
+        }
+        System.out.println(asyncInterface);
+    }
+
+    private Type makeAsyncReturnType(Type type){
+        if( type.isVoidType() ){
+            type = PrimitiveType.booleanType();
+        }
+        return new ClassOrInterfaceType(null, new SimpleName("CompletableFuture"), new NodeList<>(type));
+    }
+
+    private Set<Signature> findMissingSigs(Set<Signature> sigs, Set<Signature> expected){
+        expected = new HashSet<>(expected);
+        expected.removeAll(sigs);
+        return expected;
+    }
+
+    private Map<Signature, MethodDeclaration> methodsToSigMap(List<MethodDeclaration> methods){
+        return methods.stream()
+                .collect(toMap(
+                        MethodDeclaration::getSignature,
+                        m -> m,
+                        (e1, e2) -> e2,
+                        LinkedHashMap::new));
     }
 
     private void syncMock() throws Exception {
@@ -46,21 +105,21 @@ public class Main {
             CompilationUnit persistUnit = parseSource(getPersistPath(persist));
             Map<Signature, MethodDeclaration> sigMap =
                     listInterfaceMethods(persistUnit, persist).stream()
-                            .collect(Collectors.toMap(MethodDeclaration::getSignature, m -> m));
+                            .collect(toMap(MethodDeclaration::getSignature, m -> m));
             String mockName = getMockName(persist);
             CompilationUnit mockUnit = parseSource(getPersistMockPath(mockName));
             ClassOrInterfaceDeclaration mockClass = getClass(mockUnit, mockName);
             Map<Signature, MethodDeclaration> mockMap =
                     mockClass.getMethods().stream()
-                            .collect(Collectors.toMap(MethodDeclaration::getSignature, m -> m));
+                            .collect(toMap(MethodDeclaration::getSignature, m -> m));
             Set<Signature> missingSigs = new HashSet<>(sigMap.keySet());
             missingSigs.removeAll(mockMap.keySet());
-            if( missingSigs.size() > 0 ) {
+            if (missingSigs.size() > 0) {
                 for (Signature sig : missingSigs) {
                     MethodDeclaration method = sigMap.get(sig);
                     MethodDeclaration m = mockClass.addMethod(method.getNameAsString(), Keyword.PUBLIC);
                     m.setType(method.getType());
-                    for(Parameter param: method.getParameters()){
+                    for (Parameter param : method.getParameters()) {
                         m.addParameter(param);
                     }
                     m.addAnnotation("Override");
@@ -69,8 +128,12 @@ public class Main {
                     m.setBody(stmt);
                     System.out.printf("mock:+: %s: %s\n", mockName, m.getNameAsString());
                 }
-                Files.write(getPersistMockPath(mockName), mockUnit.toString().getBytes());
-                System.out.printf("mock:save:%s\n", getPersistMockPath(mockName).toString());
+                if (dryRun) {
+                    System.out.println(mockUnit);
+                } else {
+                    Files.write(getPersistMockPath(mockName), mockUnit.toString().getBytes());
+                    System.out.printf("mock:save:%s\n", getPersistMockPath(mockName).toString());
+                }
             }
         }
     }
@@ -116,6 +179,11 @@ public class Main {
     private ClassOrInterfaceDeclaration getClass(CompilationUnit unit, String className) {
         return unit.getClassByName(className).orElseThrow(() ->
                 new RuntimeException("Cannot find class: " + className));
+    }
+
+    private ClassOrInterfaceDeclaration getInterface(CompilationUnit unit, String interfaceName) {
+        return unit.getInterfaceByName(interfaceName).orElseThrow(() ->
+                new RuntimeException("Cannot find interface: " + interfaceName));
     }
 
     private Set<String> listMissingMockPersists() {
